+++ src/main/java/qna/domain/answer/Answer.java
@@ -1,10 +1,15 @@
-package qna.domain;
+package qna.domain.answer;
 
 import java.time.LocalDateTime;
 
 import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
+import qna.domain.AbstractEntity;
+import qna.domain.ContentType;
+import qna.domain.history.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.user.User;
 
 import javax.persistence.*;
