+++ src/main/java/qna/domain/answer/AnswerRepository.java
@@ -1,9 +1,11 @@
-package qna.domain;
+package qna.domain.answer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
+
+import qna.domain.question.Question;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
     List<Answer> findByQuestionAndDeletedFalse(Question question);
