+++ src/test/java/qna/domain/AnswerTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import qna.CannotDeleteException;
+import qna.domain.answer.Answer;
+import qna.domain.history.DeleteHistory;
 
 public class AnswerTest {
 	public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
