+++ src/test/java/qna/domain/QuestionTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import qna.CannotDeleteException;
+import qna.domain.history.DeleteHistory;
+import qna.domain.question.Question;
 
 public class QuestionTest {
 	public static final Question Q1 = new Question("title1", "contents1").writeBy(UserTest.JAVAJIGI);
